third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-4.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/sources-test.js"></script>

<script>

async function test() {
  var test = SourcesTestRunner.testJavascriptOutline;

  TestRunner.markStep('testAsyncFunction');
  await test('async function foo() { }');

  TestRunner.markStep('testAsyncArrowFunction');
  await test('var sum = async (x, y) => x + y;');

  TestRunner.markStep('testGeneratorFunction');
  await test('function* foo() { }');

  TestRunner.markStep('testMismatchBrackets');
  await test(`
function foo(a, b) {
    if (a > b) {
        return a;
}

function bar(eee) {
    foo(eee, 2 * eee);
}
`);

  TestRunner.markStep('testSyntaxError');
  await test(`
function notAGenerator(a, b) {
    yield 10;
}`);

  TestRunner.completeTest();
}

</script>

</head>

<body onload="runTest()">
<p>Verify javascript outline</p>
</body>
</html>
