third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-4-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verify javascript outline


Running: testAsyncFunction
Text:
    async function foo() { }
Outline:
    async foo():0:15

Running: testAsyncArrowFunction
Text:
    var sum = async (x, y) => x + y;
Outline:
    async sum(x, y):0:4

Running: testGeneratorFunction
Text:
    function* foo() { }
Outline:
    *foo():0:10

Running: testMismatchBrackets
Text:
    
    function foo(a, b) {
        if (a > b) {
            return a;
    }
    
    function bar(eee) {
        foo(eee, 2 * eee);
    }
    
Outline:
    foo(a, b):1:9
    bar(eee):6:9

Running: testSyntaxError
Text:
    
    function notAGenerator(a, b) {
        yield 10;
    }
Outline:
    notAGenerator(a, b):1:9
